+++ HttpFs.IntegrationTests/AssemblyInfo.fs
@@ -4,16 +4,16 @@ open System.Reflection
 
 [<assembly: AssemblyTitleAttribute("HttpFs.IntegrationTests")>]
 [<assembly: AssemblyProductAttribute("HttpFs.IntegrationTests")>]
-[<assembly: AssemblyCopyrightAttribute("Copyright © 2019")>]
+[<assembly: AssemblyCopyrightAttribute("Copyright © 2020")>]
 [<assembly: AssemblyDescriptionAttribute("A simple, functional HTTP client library for F#")>]
-[<assembly: AssemblyVersionAttribute("5.3.0")>]
-[<assembly: AssemblyFileVersionAttribute("5.3.0")>]
+[<assembly: AssemblyVersionAttribute("5.4.0")>]
+[<assembly: AssemblyFileVersionAttribute("5.4.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "HttpFs.IntegrationTests"
     let [<Literal>] AssemblyProduct = "HttpFs.IntegrationTests"
-    let [<Literal>] AssemblyCopyright = "Copyright © 2019"
+    let [<Literal>] AssemblyCopyright = "Copyright © 2020"
     let [<Literal>] AssemblyDescription = "A simple, functional HTTP client library for F#"
-    let [<Literal>] AssemblyVersion = "5.3.0"
-    let [<Literal>] AssemblyFileVersion = "5.3.0"
+    let [<Literal>] AssemblyVersion = "5.4.0"
+    let [<Literal>] AssemblyFileVersion = "5.4.0"